LPS-189642 Remove service from TaskNodeExecutor, create TimerExecutor, apply to usages and fix integration test
