+++ modules/apps/portal-workflow/portal-workflow-kaleo-runtime-impl/src/main/java/com/liferay/portal/workflow/kaleo/runtime/internal/node/TaskNodeExecutor.java
@@ -7,7 +7,6 @@ package com.liferay.portal.workflow.kaleo.runtime.internal.node;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.workflow.kaleo.definition.DelayDuration;
 import com.liferay.portal.workflow.kaleo.definition.DurationScale;
@@ -17,16 +16,13 @@ import com.liferay.portal.workflow.kaleo.definition.exception.KaleoDefinitionVal
 import com.liferay.portal.workflow.kaleo.model.KaleoInstanceToken;
 import com.liferay.portal.workflow.kaleo.model.KaleoNode;
 import com.liferay.portal.workflow.kaleo.model.KaleoTask;
-import com.liferay.portal.workflow.kaleo.model.KaleoTaskAssignment;
 import com.liferay.portal.workflow.kaleo.model.KaleoTaskInstanceToken;
 import com.liferay.portal.workflow.kaleo.model.KaleoTimer;
-import com.liferay.portal.workflow.kaleo.model.KaleoTimerInstanceToken;
 import com.liferay.portal.workflow.kaleo.model.KaleoTransition;
 import com.liferay.portal.workflow.kaleo.runtime.ExecutionContext;
 import com.liferay.portal.workflow.kaleo.runtime.assignment.AggregateKaleoTaskAssignmentSelector;
 import com.liferay.portal.workflow.kaleo.runtime.calendar.DueDateCalculator;
 import com.liferay.portal.workflow.kaleo.runtime.graph.PathElement;
-import com.liferay.portal.workflow.kaleo.runtime.internal.assignment.helper.TaskAssignerHelper;
 import com.liferay.portal.workflow.kaleo.runtime.node.BaseNodeExecutor;
 import com.liferay.portal.workflow.kaleo.runtime.node.NodeExecutor;
 import com.liferay.portal.workflow.kaleo.service.KaleoLogLocalService;
@@ -46,39 +42,8 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Michael C. Han
  */
-@Component(service = {NodeExecutor.class, TaskNodeExecutor.class})
+@Component(service = NodeExecutor.class)
 public class TaskNodeExecutor extends BaseNodeExecutor {
-
-	public void executeTimer(ExecutionContext executionContext)
-		throws PortalException {
-
-		KaleoTimerInstanceToken kaleoTimerInstanceToken =
-			executionContext.getKaleoTimerInstanceToken();
-
-		KaleoTimer kaleoTimer = kaleoTimerInstanceToken.getKaleoTimer();
-
-		kaleoActionExecutor.executeKaleoActions(
-			KaleoTimer.class.getName(), kaleoTimer.getKaleoTimerId(),
-			ExecutionType.ON_TIMER, executionContext);
-
-		List<KaleoTaskAssignment> kaleoTaskReassignments =
-			kaleoTimer.getKaleoTaskReassignments();
-
-		if (ListUtil.isNotEmpty(kaleoTaskReassignments)) {
-			_taskAssignerHelper.reassignKaleoTask(
-				kaleoTaskReassignments, executionContext);
-		}
-
-		notificationHelper.sendKaleoNotifications(
-			KaleoTimer.class.getName(), kaleoTimer.getKaleoTimerId(),
-			ExecutionType.ON_TIMER, executionContext);
-
-		if (!kaleoTimer.isRecurring()) {
-			kaleoTimerInstanceTokenLocalService.completeKaleoTimerInstanceToken(
-				kaleoTimerInstanceToken.getKaleoTimerInstanceTokenId(),
-				executionContext.getServiceContext());
-		}
-	}
 
 	@Override
 	public NodeType getNodeType() {
@@ -224,8 +189,5 @@ public class TaskNodeExecutor extends BaseNodeExecutor {
 
 	@Reference
 	private KaleoTaskLocalService _kaleoTaskLocalService;
-
-	@Reference
-	private TaskAssignerHelper _taskAssignerHelper;
 
 }
